Introduce method-style invocation to Vend::URLPatterns::Registry.
The standard subroutine invocation style for Vend::URLPattern is a bit jarring when the rest of this URLPattern stuff is all OOPy.  Particularly since the registry module is not an exporter.

Given that one works with the registry as if it were an encapsulated "object", OOPy semantics seem preferable.  Adjusted all the methods and the corresponding unit test appropriately.

Also adjusted the registration functionality to be more consistent with that of Vend::URLPatterns, and to return the number of patterns registered rather than always returning 1.