Cleaned up interface and design of Vend::Action::Factory.
Renamed the instantiate method to "create_action", which is more what it's about.

Added support for named parameters within the pattern.

Renamed _action_class_for to "action_class_for" (basically meaning it's now public), since this is potentially relevant beyond the innards of this class.

Cleaned up arguments for the create_action method to be semantically simpler and clearer.

Reworked a number of test cases in the unit test to address above changes, and added new cases to test new behavior and extant behavior more rigorously.